Fix CI build errors
