+++ src/common/mod.rs
@@ -11,6 +11,46 @@ pub(crate) mod system;
 #[cfg(feature = "user")]
 pub(crate) mod user;
 
+/// Type containing read and written bytes.
+///
+/// It is returned by [`Process::disk_usage`][crate::Process::disk_usage] and [`Disk::usage`][crate::Disk::usage].
+///
+#[cfg_attr(not(all(feature = "system", feature = "disk")), doc = "```ignore")]
+/// ```no_run
+/// use sysinfo::{Disks, System};
+///
+/// let s = System::new_all();
+/// for (pid, process) in s.processes() {
+///     let disk_usage = process.disk_usage();
+///     println!("[{}] read bytes   : new/total => {}/{} B",
+///         pid,
+///         disk_usage.read_bytes,
+///         disk_usage.total_read_bytes,
+///     );
+///     println!("[{}] written bytes: new/total => {}/{} B",
+///         pid,
+///         disk_usage.written_bytes,
+///         disk_usage.total_written_bytes,
+///     );
+/// }
+///
+/// let disks = Disks::new_with_refreshed_list();
+/// for disk in disks.list() {
+///     println!("[{:?}] disk usage: {:?}", disk.name(), disk.usage());
+/// }
+/// ```
+#[derive(Debug, Default, Clone, Copy, PartialEq, Eq, PartialOrd)]
+pub struct DiskUsage {
+    /// Total number of written bytes.
+    pub total_written_bytes: u64,
+    /// Number of written bytes since the last refresh.
+    pub written_bytes: u64,
+    /// Total number of read bytes.
+    pub total_read_bytes: u64,
+    /// Number of read bytes since the last refresh.
+    pub read_bytes: u64,
+}
+
 macro_rules! xid {
     ($(#[$outer:meta])+ $name:ident, $type:ty $(, $trait:ty)?) => {
         #[cfg(any(feature = "system", feature = "user"))]
